// src/types.ts
export interface Account {
    account_category: string;
    account_code: string;
    account_currency: string;
    account_identifier: string;
    account_status: string;
    account_name: string;
    account_type: string;
    account_type_bank: string;
    system_account: string;
    value_type: string;
    total_value: number;
};